make index integer

+++ src/1_pop_data/10_create_master_index.R
@@ -60,7 +60,7 @@ time_index_expanded <- tibble(
   mutate(
     t_name = floor_date(as.Date(collection_date), "week", week_start = 1),
     t_key = str_replace_all(as.character(t_name), "-", "") |> as.integer(),
-    t = t_name |> as.character() |> as_factor() |> as.numeric()
+    t = t_name |> as.character() |> as_factor() |> as.integer()
   )
 
 time_index <- time_index_expanded |>
@@ -81,7 +81,7 @@ agesex_index <- lapply(
     a = as.integer(factor(paste0(age_min, age_max))),
     a_key = paste0(age_min, str_pad(age_max, 3, pad = "0")) |> as.integer(),
     s_name = str_sub(agesex, 1, 1),
-    s = s_name |> as_factor() |> as.numeric(),
+    s = s_name |> as_factor() |> as.integer(),
     s_key = s
   ) |>
   select(-age_min, -age_max, -gender)
